chore: organise types and mock data files and add new variable to enviroment
