+++ src/environments/environment.prod.ts
@@ -1,4 +1,5 @@
 export const environment = {
   production: true,
   apiUrl: 'https://pokeapi.co/api/v2',
+  pokemonsPerPage: 12,
 };
