+++ src/environments/environment.ts
@@ -5,6 +5,7 @@
 export const environment = {
   production: false,
   apiUrl: 'https://pokeapi.co/api/v2',
+  pokemonsPerPage: 12,
 };
 
 /*
